+++ lizard_map/urls.py
@@ -131,7 +131,7 @@ urlpatterns = patterns(
      'lizard_map.map_location_load_default'),
 
     # Download map as image
-    (r'^download/',
+    (r'^download-map/',
      'lizard_map.views.save_map_as_image',
      {},
      'lizard_map.views.save_map_as_image'),